Created the directory structure for the project and a skeleton app.
The directory structure for the project has been created. The
pasteapp folder contains the actual application templates and views.
The top level directory will contain the configuration and management
scripts.

There will eventually also be a tests folder that will
contain the unit test scripts for the application. Currently there are
no unit tests in place - the focus for this first commit was creating
the skeleton application to ensure that the environment has been set up
correctly.

The application can be run by going into the top level directory and
executing:

python manage.py runserver.

This will start a development server on localhost:5000. The application
can then be viewed by opening a browser window and going to:

http://localhost:5000