Added configuration and outline for the database models.
The application can now be configured for development or testing.
A third configuration file - for production still needs to be created.

In addition, basic models have also been created and the database
functions have also been added.

The outline of the unit tests have not been included in this commit.